fixed code.
"И где самое главное - КТО СОВЕРШИЛ ЭТО ДЕЙСТВИЕ?"
This comment didn't quite understand what needs to be done?
