--- src/main/java/faang/school/accountservice/controller/BalanceAuditController.java
@@ -1,13 +1,0 @@
-package faang.school.accountservice.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/balance_audit")
-@RequiredArgsConstructor
-public class BalanceAuditController {
-
-
-}
